player code rebase

+++ bot.py
@@ -18,25 +18,27 @@ with open("credentials.json", "r") as creds:
 
 client = commands.Bot(command_prefix='-')  # prefix our commands with '-'
 
-# players = {}
 masters = {}
 queues = {}
 queuelocks = {}
 
+
+
 # main queue manager function to play music in queues
-async def check_queue(id, voice, ctx, msg):
+async def check_queue(id, voice, ctx, msg=None):
     while voice.is_playing() or voice.is_paused():
         await asyncio.sleep(5)
-    await msg.delete()
+    if msg:
+        await msg.delete()
     if queues[id] != []:
-        # print(queues[id][0]["URL"])
         embed=discord.Embed(title="Currently Playing", description=f'[{queues[id][0]["title"]}]({queues[id][0]["url"]})', color=0xfe4b81)
         embed.set_thumbnail(url=queues[id][0]["thumbnails"][len(queues[id][0]["thumbnails"])-1]["url"])
         voice.play(FFmpegPCMAudio(queues[id][0]["link"], **FFMPEG_OPTIONS))
-        # await ctx.send('Currently playing '+queues[id][0]["title"])
         msg = await ctx.send(embed=embed)
         queues[id].pop(0)
         await check_queue(id, voice, ctx, msg)
+
+
 
 # check if bot is ready
 @client.event  
@@ -76,7 +78,6 @@ async def play(ctx, *,keyw):
 
     voice = get(client.voice_clients, guild=ctx.guild)
 
-    # print(voice.is_playing())
     try:
         if voice:
             if not masters[ctx.guild.id].voice or masters[ctx.guild.id].voice.channel != voice.channel or (not (voice.is_playing() or voice.is_paused()) and queues[ctx.guild.id] == []):
@@ -85,16 +86,20 @@ async def play(ctx, *,keyw):
             if not (voice.is_playing() or voice.is_paused()) and queues[ctx.guild.id] == []:
                 with YoutubeDL(YDL_OPTIONS) as ydl:
                     info = ydl.extract_info(url, download=False)
-                URL = info['url']
-                # for song in queues[ctx.guild.id]:
-                # print(URL)
-                embed=discord.Embed(title="Currently playing", description=f'[{info["title"]}]({url})', color=0xfe4b81)
-                embed.set_thumbnail(url=info["thumbnails"][len(info["thumbnails"])-1]["url"])
-                voice.play(FFmpegPCMAudio(URL, **FFMPEG_OPTIONS))
-                if voice.is_playing():
-                    # await ctx.send('Currently playing '+info["title"])
-                    msg = await ctx.send(embed=embed)
-                await check_queue(ctx.guild.id, voice, ctx, msg)
+                data = {
+                    "link": info['url'],
+                    "url": url,
+                    "title": info['title'],
+                    "thumbnails": info["thumbnails"],
+                    "raw": info
+                }
+                queues[ctx.guild.id].append(data)
+                # embed=discord.Embed(title="Currently playing", description=f'[{info["title"]}]({url})', color=0xfe4b81)
+                # embed.set_thumbnail(url=info["thumbnails"][len(info["thumbnails"])-1]["url"])
+                # voice.play(FFmpegPCMAudio(URL, **FFMPEG_OPTIONS))
+                # if voice.is_playing():
+                #     msg = await ctx.send(embed=embed)
+                await check_queue(ctx.guild.id, voice, ctx)
 
             else:
                 with YoutubeDL(YDL_OPTIONS) as ydl:
@@ -103,42 +108,43 @@ async def play(ctx, *,keyw):
                     "link": info['url'],
                     "url": url,
                     "title": info['title'],
-                    "thumbnails": info["thumbnails"]
+                    "thumbnails": info["thumbnails"],
+                    "raw": info
                 }
                 queues[ctx.guild.id].append(data)
-                # print(ctx.guild.id)
-                # print(queues[ctx.guild.id])
                 embed=discord.Embed(title="Item queued", description=f'[{info["title"]}]({url})', color=0xfe4b81)
                 embed.set_thumbnail(url=info["thumbnails"][len(info["thumbnails"])-1]["url"])
-                # await ctx.send("Item queued "+info["title"])
                 await ctx.send(embed=embed)
-                return
         else: 
             if ctx.message.author.voice:
                 channel = ctx.message.author.voice.channel
-                # voice = get(client.voice_clients, guild=ctx.guild)
                 voice = await channel.connect()
                 masters[ctx.guild.id] = ctx.message.author
                 queues[ctx.guild.id] = []
                 with YoutubeDL(YDL_OPTIONS) as ydl:
                     info = ydl.extract_info(url, download=False)
-                URL = info['url']
-                # for song in queues[ctx.guild.id]:
-                # print()
-                embed=discord.Embed(title="Currently playing", description=f'[{info["title"]}]({url})', color=0xfe4b81)
-                embed.set_thumbnail(url=info["thumbnails"][len(info["thumbnails"])-1]["url"])
-                voice.play(FFmpegPCMAudio(URL, **FFMPEG_OPTIONS))
-                if voice.is_playing():
-                    # await ctx.send('Currently playing '+info["title"])
-                    msg = await ctx.send(embed=embed)
-                await check_queue(ctx.guild.id, voice, ctx, msg)
+                data = {
+                    "link": info['url'],
+                    "url": url,
+                    "title": info['title'],
+                    "thumbnails": info["thumbnails"],
+                    "raw": info
+                }
+                queues[ctx.guild.id].append(data)
+                # embed=discord.Embed(title="Currently playing", description=f'[{info["title"]}]({url})', color=0xfe4b81)
+                # embed.set_thumbnail(url=info["thumbnails"][len(info["thumbnails"])-1]["url"])
+                # voice.play(FFmpegPCMAudio(URL, **FFMPEG_OPTIONS))
+                # if voice.is_playing():
+                #     msg = await ctx.send(embed=embed)
+                await check_queue(ctx.guild.id, voice, ctx)
             else:
                 embed=discord.Embed(title="You are currently not connected to any voice channel", color=0xfe4b81)
-                # await ctx.send('You are not currently connected to any voice channel')
                 await ctx.send(embed=embed, delete_after=10)
     except:
         embed=discord.Embed(title="can't play the requested audio", color=0xfe4b81)
         await ctx.send(embed=embed, delete_after=10)
+
+
 
 # shows the queued songs of the ctx guild
 @client.command(name="queue")
@@ -152,6 +158,8 @@ async def listQueue(ctx):
         out = "None"
     embed=discord.Embed(title="Currently in queue", description=out, color=0xfe4b81)
     await ctx.send(embed=embed)
+
+
 
 # removes a mentioned song from queue and displays it
 @client.command(name="remove")
@@ -169,6 +177,8 @@ async def removeQueueSong(ctx, index: int):
     else:
         embed=discord.Embed(title="Invalid request", color=0xfe4b81)
     await ctx.send(embed=embed)
+
+
 
 # command to resume voice if it is paused
 @client.command()
@@ -227,6 +237,8 @@ async def skip(ctx):
         embed=discord.Embed(title="I am currently not connected to any voice channel", color=0xfe4b81)
         await ctx.send(embed=embed, delete_after=7)
 
+
+
 # stops the bot player by clearing the current queue and skipping the current audio
 @client.command()
 async def stop(ctx):
@@ -248,6 +260,8 @@ async def stop(ctx):
         embed=discord.Embed(title="I am currently not connected to any voice channel", color=0xfe4b81)
         await ctx.send(embed=embed, delete_after=7)
 
+
+
 # leaves the vc on demand
 @client.command(name='leave', help='To make the bot leave the voice channel')
 async def leave(ctx):
@@ -265,12 +279,16 @@ async def leave(ctx):
         embed=discord.Embed(title="I am currently not connected to a voice channel.", color=0xfe4b81)
         await ctx.send(embed=embed, delete_after=7)
 
+
+
 # command to clear channel messages
 @client.command()
 async def clear(ctx, amount=5):
     await ctx.channel.purge(limit=amount)
     embed=discord.Embed(title="Messages have been cleared", color=0xfe4b81)
     await ctx.send(embed=embed, delete_after=7)
+
+
 
 @client.command()
 async def lock(ctx):
